+++ implementations/c/ziggurat.c
@@ -7,7 +7,7 @@
 
 /******************************************************************************/
 
-float r4_exp ( unsigned long int *jsr, int ke[256], float fe[256],
+float r4_exp ( unsigned int *jsr, int ke[256], float fe[256],
 		float we[256] )
 
 	/******************************************************************************/
@@ -161,7 +161,7 @@ void r4_exp_setup ( int ke[256], float fe[256], float we[256] )
 }
 /******************************************************************************/
 
-float r4_nor ( unsigned long int *jsr, int kn[128], float fn[128],
+float r4_nor ( unsigned int *jsr, int kn[128], float fn[128],
 		float wn[128] )
 
 	/******************************************************************************/
@@ -338,7 +338,7 @@ void r4_nor_setup ( int kn[128], float fn[128], float wn[128] )
 }
 /******************************************************************************/
 
-float r4_uni ( unsigned long int *jsr )
+float r4_uni ( unsigned int *jsr )
 
 	/******************************************************************************/
 	/*
@@ -373,7 +373,7 @@ float r4_uni ( unsigned long int *jsr )
 	   the range [0,1].
 	 */
 {
-	unsigned long int jsr_input;
+	unsigned int jsr_input;
 	float value;
 
 	jsr_input = *jsr;
@@ -388,7 +388,7 @@ float r4_uni ( unsigned long int *jsr )
 }
 /******************************************************************************/
 
-unsigned long int shr3 ( unsigned long int *jsr )
+unsigned int shr3 ( unsigned int *jsr )
 
 	/******************************************************************************/
 	/*
@@ -423,7 +423,7 @@ unsigned long int shr3 ( unsigned long int *jsr )
 	   Output, unsigned long int SHR3, the new value.
 	 */
 {
-	unsigned long int value;
+	unsigned int value;
 
 	value = *jsr;
